Add files via upload
In this version, the passive cable properties were slightly tuned to capture the distance-dependent space clamp effect measured in an experiment (Ref. [Williams & Mitchell, Nature Neuroscience, 2008]).  In addition, the dynamics  of synaptic conductance has been revised to be the difference between two exponential functions in order to investigate how the performance of SIM and IM depends on the time course of the synaptic inputs.